refactor: removed the admin and shared folders
