+++ app/Http/Requests/V1/Authentication/OnboardingRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests\V1\Shared\Authentication;
+namespace App\Http\Requests\V1\Authentication;
 
 use Illuminate\Foundation\Http\FormRequest;
