--- routes/Api/V1/Admin/index.php
@@ -1,5 +1,0 @@
-<?php
-
-use Illuminate\Support\Facades\Route;
-
-Route::prefix('user-management')->group(__DIR__.'/UserManagementRoutes.php');
