+++ routes/Api/V1/AuthenticationRoutes.php
@@ -1,7 +1,7 @@
 <?php
 
-use App\Http\Controllers\V1\Shared\Authentication\LoginController;
-use App\Http\Controllers\V1\Shared\Authentication\OnboardingController;
+use App\Http\Controllers\V1\Authentication\LoginController;
+use App\Http\Controllers\V1\Authentication\OnboardingController;
 use Illuminate\Support\Facades\Route;
 
 Route::post('/register', OnboardingController::class);
