--- routes/Api/V1/Shared/index.php
@@ -1,5 +1,0 @@
-<?php
-
-use Illuminate\Support\Facades\Route;
-
-Route::prefix('authentication')->group(__DIR__.'/AuthenticationRoutes.php');
